Major visual changes to front end of website
Use of Bootstrap Framework to add:
- Navigation Bar
- Jumbotron Title
- Visually appealing carousel

Editing of the existing CSS file: It is now almost unused as Bootstrap takes care of a lot of components. Still present for the position and size of map.

Adding of a tab icon wifind.ico, changes seen in the <head> section of master-webpage.html.